Add files via upload
This Assessor has a better button UI on state 3 and responds to the check boxes being checked. It also responds to the <enter> key being pressed on state 4. I did not get around to fixing the initial menu dialog although I will be on the lookout for a solution to that within the next 24 hours.